refactor: extract lib.rs and move server wiring into app module
Introduce a library crate with app::run as the composition root so main.rs
only handles CLI parsing, telemetry init, and logging setup.
